# lib/api-spec/openapi.yaml
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ReviewDecision:
      properties:
        decision:
          pattern: ^(promote|reject)$
          title: Decision
          type: string
        note:
          default: ''
          maxLength: 500
          title: Note
          type: string
        source:
          maxLength: 500
          minLength: 1
          title: Source
          type: string
        target:
          maxLength: 500
          minLength: 1
          title: Target
          type: string
        tier:
          default: ''
          pattern: ^(r\d+\.\w+)?$
          title: Tier
          type: string
        type:
          maxLength: 60
          minLength: 1
          title: Type
          type: string
      required:
      - source
      - type
      - target
      - decision
      title: ReviewDecision
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Ingest repositories and explore their structure as a typed knowledge graph
  title: Api
  version: 2.0.0
openapi: 3.1.0
paths:
...
  /api/v2/code-bridges:
    get:
      description: 'Where the selected repositories actually touch, at code altitude.


        The per-repo graph is deliberately intra-repo: files, classes, endpoints of

        ONE codebase. Drawing several of those together produces disconnected

        islands, because nothing in that response crosses a repository boundary.


        What crosses is a rendezvous: a call site INVOKES a contract that another

        repo''s endpoint EXPOSES. The contract is the join, so it is returned as a

        node and both halves as edges.


        The two endpoint nodes come back as well. They are the whole point of the

        query, and the caller''s per-repo sample is capped -- a bridge whose call

        site fell outside that sample would otherwise dangle.'
      operationId: code_bridges_api_v2_code_bridges_get
      parameters:
      - description: Comma-separated repo ids
        in: query
        name: repos
        required: true
        schema:
          description: Comma-separated repo ids
          title: Repos
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 200
          maximum: 500
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Code Bridges
  /api/v2/config/ownership:
    get:
      description: "VQ4: who defines a config/env key, who reads it, and which teams own\nthe repos on\
        \ each side. Runs over claims directly \u2014 definition and read\nsites are facts even when no\
        \ resolver joined them."
      operationId: config_ownership_api_v2_config_ownership_get
      parameters:
      - in: query
        name: name
        required: true
        schema:
          maxLength: 200
          minLength: 1
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Config Ownership
  /api/v2/impact/library:
    get:
      description: 'VQ2: who breaks when this library changes.


        `key` is a version-free purl (`pkg:maven/org.acme/billing-lib`). Version

        skew falls out of the per-edge versions on one rendezvous node.'
      operationId: library_impact_api_v2_impact_library_get
      parameters:
      - in: query
        name: key
        required: true
        schema:
          maxLength: 300
          minLength: 5
          title: Key
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Library Impact
  /api/v2/impact/repo/{repo_id}:
    get:
      description: 'VQ2 from the repo side: what this repo publishes and who consumes it.'
      operationId: repo_impact_api_v2_impact_repo__repo_id__get
      parameters:
      - in: path
        name: repo_id
        required: true
        schema:
          title: Repo Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Repo Impact
...
  /api/v2/links/deprecated:
    get:
      description: "Deprecated contracts, and exposed contracts no one invokes\n(dead-endpoint candidates\
        \ \u2014 absence of a consumer is a candidate fact,\nnot proof, so they are reported, never deleted)."
      operationId: deprecated_and_dead_api_v2_links_deprecated_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 200
          maximum: 500
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Deprecated And Dead
...
  /api/v2/links/review:
    get:
      description: "Candidate edges awaiting review, plus every decline counter\nfrom the latest run \u2014\
        \ the counters ARE the missing-recall worklist."
      operationId: review_queue_api_v2_links_review_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 500
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Review Queue
  /api/v2/links/review/decide:
    post:
      description: 'Apply an operator decision now AND persist it to the control plane so

        every future relink replays it (D6 promotions replay).'
      operationId: review_decide_api_v2_links_review_decide_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDecision'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Review Decide
...
  /api/v2/service-map:
    get:
      operationId: service_map_api_v2_service_map_get
      parameters:
      - in: query
        name: min_confidence
        required: false
        schema:
          default: 0.6
          maximum: 1.0
          minimum: 0.0
          title: Min Confidence
          type: number
      - in: query
        name: limit
        required: false
        schema:
          default: 500
          maximum: 500
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Service Map
...